+++ ext/TimeVaryingInputs0DExt.jl
@@ -17,7 +17,7 @@ import ClimaUtilities.TimeVaryingInputs:
     Flat,
     PeriodicCalendar,
     extrapolation_bc
-import ClimaUtilities.TimeManager: ITime, date
+import ClimaUtilities.TimeManager: ITime
 
 """
     InterpolatingTimeVaryingInput0D
@@ -193,41 +193,30 @@ end
 Check if the given `time` is in the range of definition for `itp`.
 """
 function Base.in(time, itp::InterpolatingTimeVaryingInput0D)
+    time = _normalize_time(itp.range, time)
     return itp.range[1] <= time <= itp.range[2]
 end
 
 """
-    _normalize_time(itp::InterpolatingTimeVaryingInput0D, time)
+    _normalize_time(range::Tuple, time)
 
-Convert `time` to the time type used by `itp.times`.
+Convert `time` to the type of the times in `range`, the first and last times of
+a time varying input.
 """
-_normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:Number}},
-    time::Number,
-) = time
-_normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:Number}},
-    time::ITime,
-) = eltype(itp.range)(float(time))
-_normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:Number}},
-    time::DateTime,
-) = error(
+_normalize_time(range::Tuple{<:Number, <:Number}, time::Number) = time
+_normalize_time(range::Tuple{<:Number, <:Number}, time::ITime) =
+    eltype(range)(float(time))
+_normalize_time(range::Tuple{<:Number, <:Number}, time::DateTime) = error(
     "Cannot evaluate InterpolatingTimeVaryingInput0D with times as numbers and inputs as DateTime",
 )
-_normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:ITime}},
-    time::ITime,
-) = time
-_normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:ITime}},
-    time::Number,
-) = first(promote(ITime(time), itp.range[1]))
-function _normalize_time(
-    itp::InterpolatingTimeVaryingInput0D{<:AbstractArray{<:ITime}},
-    time::DateTime,
-)
-    epoch = date(itp.range[1])
+_normalize_time(range::Tuple{<:ITime, <:ITime}, time::ITime) =
+    first(promote(time, range[1]))
+_normalize_time(range::Tuple{<:ITime, <:ITime}, time::Number) =
+    first(promote(ITime(time), range[1]))
+function _normalize_time(range::Tuple{<:ITime, <:ITime}, time::DateTime)
+    epoch = range[1].epoch
+    isnothing(epoch) &&
+        error("Cannot evaluate at a DateTime when the times have no epoch")
     elapsed = time - epoch
     return ITime(elapsed.value; period = typeof(elapsed)(1), epoch)
 end
@@ -324,7 +313,9 @@ end
 
 @inline function _gap_stencil(time, times, t_end, dt, ::NearestNeighbor)
     w = _zero_weight(time, times)
-    time >= t_end + 0.5dt && return (firstindex(times), firstindex(times), w)
+    # Scaling an ITime by a float rounds its counter, so use
+    # 2 * (time - t_end) >= dt instead of time >= t_end + 0.5dt
+    2 * (time - t_end) >= dt && return (firstindex(times), firstindex(times), w)
     return (lastindex(times), lastindex(times), w)
 end
 
@@ -354,7 +345,7 @@ function TimeVaryingInputs.evaluate!(
     args...;
     kwargs...,
 )
-    return _evaluate!(dest, itp, _normalize_time(itp, time))
+    return _evaluate!(dest, itp, _normalize_time(itp.range, time))
 end
 
 """